kdrive: Add a proper header for the linux drivers
Stop declaring `extern` variables/functions throught the code.
Move ddxInit out of the linux driver code.
